# Body parsing optimization